add swarm-agnostic swarms, browser library
