+++ .gitignore
@@ -1,4 +1,5 @@
-/node_modules
+node_modules
 /.data
+/dist
+/repos.json
 
-repos.json
