+++ lib/distribution.js
@@ -1,5 +1,5 @@
 const {DiscoverySwarm} = require('./discovery/swarm')
-const {ResponseSwarm} = require('./bridge/response-swarm')
+const {ResponseSwarm} = require('./response/swarm')
 const debug = require('debug')('fm2u-distribution')
 
 const swarms = new Map()
