+++ lib/request/abstract-swarm.js
@@ -1,14 +1,7 @@
-const Hyperswarm = require('hyperswarm')
-const {EventEmitter} = require('events')
-const pump = require('pump')
-const crypto = require('crypto')
-const uuid = require('uuid/v1')
-const lps = require('length-prefixed-stream')
-const {RequestStream} = require('./request-stream')
+const {AbstractSwarm} = require('../abstract-swarm')
+const {RequestStream} = require('./stream')
 const {ResponseCode} = require('../messages')
 const debug = require('debug')('fm2u-request-swarm')
-
-const {ProtocolSwarm} = require('../protocol-swarm')
 
 const requestTypeMapping = {
   GET: 'OK',
@@ -19,10 +12,11 @@ const requestTypeMapping = {
 
 const responseCodes = Object.keys(ResponseCode).map(code => code.toLowerCase())
 
-class RequestSwarm extends ProtocolSwarm {
-  constructor(docUrl, opts) {
+class AbstractRequestSwarm extends AbstractSwarm {
+  constructor(docUrl, createSwarm, opts) {
     super({
       topic: `annotations-${docUrl}`,
+      createSwarm,
       lookup: true,
       announce: false,
     })
@@ -73,4 +67,4 @@ class RequestSwarm extends ProtocolSwarm {
   }
 }
 
-module.exports = {RequestSwarm}
+module.exports = {AbstractRequestSwarm}
