+++ lib/request/browser-swarm.js
@@ -1,0 +1,12 @@
+const {BrowserSwarm} = require('hyperswarm-ws')
+const {AbstractRequestSwarm} = require('./abstract-swarm')
+
+const gatewayUrl = 'ws://localhost:4200'
+
+class RequestSwarm extends AbstractRequestSwarm {
+  constructor(docUrl, opts) {
+    super(docUrl, () => new BrowserSwarm(gatewayUrl), opts)
+  }
+}
+
+module.exports = {RequestSwarm}
