+++ lib/request/stream.js
@@ -6,7 +6,7 @@ const {
   ResponseCode,
 } = require('../messages')
 
-class RequestStream extends Duplex {
+class Stream extends Duplex {
   constructor() {
     super()
     this.setMaxListeners(256)
@@ -47,4 +47,4 @@ class RequestStream extends Duplex {
   _read() {}
 }
 
-module.exports = {RequestStream}
+module.exports = {RequestStream: Stream}
