+++ lib/response/stream.js
@@ -6,7 +6,7 @@ const {
   RequestMethod,
 } = require('../messages')
 
-class ResponseStream extends Duplex {
+class Stream extends Duplex {
   constructor() {
     super()
     this.setMaxListeners(256)
@@ -47,4 +47,4 @@ class ResponseStream extends Duplex {
   _read() {}
 }
 
-module.exports = {ResponseStream}
+module.exports = {ResponseStream: Stream}
