+++ lib/response/swarm.js
@@ -1,6 +1,6 @@
 const pathMatch = require('path-match')
 const {ProtocolSwarm} = require('../protocol-swarm')
-const {ResponseStream} = require('./response-stream')
+const {ResponseStream} = require('./stream')
 const {RequestMethod} = require('../messages')
 const debug = require('debug')('fm2u-response-swarm')
 
@@ -14,7 +14,7 @@ const routeMatcher = pathMatch({
   end: false,
 })
 
-class ResponseSwarm extends ProtocolSwarm {
+class Swarm extends ProtocolSwarm {
   handlers = []
 
   constructor(id) {
@@ -71,4 +71,4 @@ class ResponseSwarm extends ProtocolSwarm {
   }
 }
 
-module.exports = {ResponseSwarm}
+module.exports = {ResponseSwarm: Swarm}
